Added logger

+++ MVC/Controllers/BooksController.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text.Json;
 using System.Threading.Tasks;
 using AutoMapper;
 using DAL.EF;
@@ -9,6 +10,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Internal;
+using Microsoft.Extensions.Logging;
 using MVC.Models;
 
 namespace MVC.Controllers
@@ -17,28 +19,50 @@ namespace MVC.Controllers
     {
         private readonly LibraryContext _db;
         private readonly IMapper _mapper;
-        public BooksController(LibraryContext context,IMapper _mapper)
+        private readonly ILogger<BooksController> _logger;
+        public BooksController(LibraryContext context,IMapper _mapper, ILogger<BooksController> logger)
         {
             this._db = context;
             this._mapper = _mapper;
+            this._logger = logger;
         }
 
         // GET: Books
         [ActionName("Index")]
         public async Task<ActionResult> Index()
         {
-            var list = await _db.Books.ToListAsync();
-            return View(_mapper.Map<IEnumerable<BookModel>>(list));
+            _logger.LogInformation("Openning Index Page");
+            try
+            {
+                var list = await _db.Books.ToListAsync();
+                return View(_mapper.Map<IEnumerable<BookModel>>(list));
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e, "Exception in Index page");
+                return View();
+            }
+            finally {
+                _logger.LogInformation("Index finished");
+            }
         }
 
         [ActionName("Get")]
         public ActionResult Get(int id) {
+            _logger.LogInformation("Get method action starts",id);
             var book = GetBookWithTags(id);
             if (book != null)
             {
-                return View(book);
+                try
+                {
+                    return View(book);
+                }
+                finally {
+                    _logger.LogInformation("Get method action successfully finished",);
+                }
             }
             else {
+                _logger.LogError("Book with such id doesn't exist");
                 return RedirectToAction("Index");
             }
         }
@@ -55,24 +79,54 @@ namespace MVC.Controllers
         // GET: Books/Create
         public ActionResult Create()
         {
-            var tags = _mapper.Map<IEnumerable<TagModel>>(_db.Tags);
-            var book = new BookModel { Tags = tags.ToList()};
-            return View(book);
+            try
+            {
+                _logger.LogInformation("Create action method starts");
+                var tags = _mapper.Map<IEnumerable<TagModel>>(_db.Tags);
+                var book = new BookModel { Tags = tags.ToList() };
+                return View(book);
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e, "Exception in Create action");
+                return RedirectToAction("Index");
+            }
+            finally {
+                _logger.LogInformation("Create action method finished");
+            }
         }
 
         [HttpGet]
 
         public ActionResult Update(int id) {
-            var book = GetBookWithTags(id);
-            var bookTags = book.Tags;
-            var tags = _mapper.Map<List<TagModel>>(_db.Tags);
-            book.Tags = tags;
-            for(int i = 0; i < tags.Count; i++) {
-                if (bookTags.Select(x=>x.Id).Contains(tags[i].Id)) {
-                    book.Tags[i].Checked = true;
+            try
+            {
+                _logger.LogInformation("Update method action starts", id);
+                var book = GetBookWithTags(id);
+                if (book == null)
+                {
+                    throw new NullReferenceException("There isn't book with such id");
                 }
+                var bookTags = book.Tags;
+                var tags = _mapper.Map<List<TagModel>>(_db.Tags);
+                book.Tags = tags;
+                for (int i = 0; i < tags.Count; i++)
+                {
+                    if (bookTags.Select(x => x.Id).Contains(tags[i].Id))
+                    {
+                        book.Tags[i].Checked = true;
+                    }
+                }
+                return View(book);
             }
-            return View(book);
+            catch (Exception e)
+            {
+                _logger.LogError(e, "Exception in Create action");
+                return RedirectToAction("Index");
+            }
+            finally {
+                _logger.LogInformation("Update method action finished", id);
+            }
         }
 
         // POST: Books/Create
@@ -81,6 +135,7 @@ namespace MVC.Controllers
         {
             try
             {
+                _logger.LogInformation("Create post method starts", JsonSerializer.Serialize(book));
                 var item = _mapper.Map<Book>(book);
                 var tags = book.Tags;
                 item.BookTags = tags
@@ -94,7 +149,12 @@ namespace MVC.Controllers
             }
             catch (Exception e)
             {
+                _logger.LogError(e,"Exception in Create post method");
                 return View("Views/Shared/Error.cshtml");
+            }
+            finally
+            {
+                _logger.LogInformation("Create post method finished");
             }
         }
 
@@ -104,25 +164,30 @@ namespace MVC.Controllers
         {
             try
             {
-                var bookFromBd = _db.Books.Include(x=>x.BookTags)
-                    .Single(x=>x.Id == book.Id);
+                _logger.LogInformation("Edit post method starts", JsonSerializer.Serialize(book));
+                var bookFromBd = _db.Books.Include(x => x.BookTags)
+                    .Single(x => x.Id == book.Id);
+                if (bookFromBd == null) 
+                {
+                    throw new NullReferenceException("There isn't such book");
+                }
                 var tags = book.Tags
                     .Where(x => x.Checked == true)
                     .Select(y => new BookTags { TagId = y.Id, BookId = book.Id })
                     .ToList();
-                for (int i = 0; i < bookFromBd.BookTags.Count(); i++) 
+                for (int i = 0; i < bookFromBd.BookTags.Count(); i++)
                 {
                     if (!book.Tags
-                        .Where(x=>x.Checked)
+                        .Where(x => x.Checked)
                         .Select(x => x.Id)
-                        .Contains(bookFromBd.BookTags.ToList()[i].TagId)) 
+                        .Contains(bookFromBd.BookTags.ToList()[i].TagId))
                     {
                         bookFromBd.BookTags.Remove(bookFromBd.BookTags.ToList()[i]);
                     }
                 }
-                for (int i = 0; i < tags.Count(); i++) 
+                for (int i = 0; i < tags.Count(); i++)
                 {
-                    if (!bookFromBd.BookTags.Select(x=>x.TagId).Contains(tags[i].TagId)) 
+                    if (!bookFromBd.BookTags.Select(x => x.TagId).Contains(tags[i].TagId))
                     {
                         bookFromBd.BookTags.Add(tags[i]);
                     }
@@ -134,9 +199,13 @@ namespace MVC.Controllers
 
                 return RedirectToAction("Index");
             }
-            catch(Exception e)
+            catch (Exception e)
             {
+                _logger.LogError(e,"Exception in edit post method",JsonSerializer.Serialize(book));
                 return View("Views/Books/Update.cshtml");
+            }
+            finally {
+                _logger.LogInformation("Edit post method finished");
             }
         }
 
@@ -146,15 +215,24 @@ namespace MVC.Controllers
         {
             try
             {
+                _logger.LogInformation("Delete post method starts", id);
                 var book = await _db.Books.FindAsync(id);
+                if (book == null)
+                {
+                    throw new NullReferenceException("There isn't such book");
+                }
                 _db.Books.Remove(book);
                 await _db.SaveChangesAsync();
 
                 return RedirectToAction("Index");
             }
-            catch
+            catch(Exception e)
             {
+                _logger.LogError(e, "Exception in delete post method");
                 return View();
+            }
+            finally {
+                _logger.LogInformation("Delete post method finished");
             }
         }
     }
